requirements.txt
numpy==1.23.5
pandas==1.5.2
PyAutoGUI==0.9.53
pokereval==0.1.0
easyocr==1.6.2
keyboard==0.13.5
pytweening==1.0.4
opencv-python==4.8.1.78
Pillow==9.5.0
pytesseract==0.3.10
matplotlib==3.7.2
PySimpleGUI==4.60.5
pyperclip==1.8.2
